Version 0.9.16.9000
Eliminate writing of correspondence table to the tempdir() in atlasMaker().

Update atlasMaker() for DRB1 in release version 3.48.0, HLA-B in releases 3.44.0 and 3.43.0, HLA-C in release 3.02.0, HFE in relases 3.27.0-3.22.0, HLA-DPA, -DPB, TAP1 and TAP2 in release 3.00.0, HLA-V in release 3.14.0, HLA-DPA2 and -DPB2 in versions 3.53.0-3.27.0, HLA-N, -S, -U, and -Y in versons 3.35.0-3.33.0, HLA-Y in versions 3.32.0-3.20.0, HLA-W and -T in version 3.27.0, for all loci in version 3.13.0/3.13.1, HLA-C in version 3.02.0, and HLA-DPA, -DPB, TAP1 and TAP2 in version 3.00.0, so that all atlases are built for all source types in all releases.

Update documentation for updateAll().